world/ folder split into 2 files
