+++ src/world/index.js
@@ -1,1 +1,2 @@
 export * from "./world.js";
+export * from "./block_type.js";
