+++ src/world/world.js
@@ -1,5 +1,8 @@
+import {fromNested, exportAs, assert} from '../utils.js';
 import {GameComponent} from '../game_component.js';
-import {fromNested, exportAs, assert, classOf, isString} from '../utils.js';
+
+import {Blocks} from './block_type.js';
+
 
 export var SIZE = [16, 16, 16];
 export var LOW = [-8, -8, -8];
@@ -63,64 +66,6 @@ export class World extends GameComponent {
     }
   }
 }
-
-
-export var Blocks = {};
-
-
-export class BlockType {
-  static BlockByNum = [];
-  static BlockByName = {};
-  
-  constructor(config){
-    let cls = classOf(this);
-    assert(config);
-    if(isString(config)){
-      config = {name: config};
-    }
-    this.config = config;
-    Object.assign(this, config);
-    // num
-    this.num = cls.BlockByNum.length;
-    // name
-    this.name = config.name ?? this._getPlaceholderName();
-    assert(cls.BlockByName[this.name] == null);
-    // other
-    this.transparent = this.config.transparent ?? false;
-    this._addToRegistry();
-  }
-
-  _getPlaceholderName(){
-    return `[[Block_${this.num}]]`
-  }
-
-  _addToRegistry(updateGlob=true){
-    let cls = classOf(this);
-    cls.BlockByName[this.name] = this;
-    cls.BlockByNum.push(this);
-    if(updateGlob){
-      cls.updateRegistry();
-    }
-  }
-
-  static updateRegistry(){
-    Object.assign(Blocks, this.BlockByName, this.BlockByNum);
-  }
-
-  static addTypes(...types){
-    return Object.fromEntries(types.map(t => {
-      let b = new BlockType(t);
-      return [b.name, b];
-    }))
-  }
-}
-
-BlockType.addTypes(
-  {name: 'air', transparent: true},
-  {name: 'grass'}
-)
-
-
 
 
 exportAs(World);
